pebble: relax tracing_subscriber setup
When multiple tests try to set up the tracing subscriber based on the
env we're happy for whichever does it first to "win" and for the other
tests to do nothing.

This commit switches the http_01 test's setup to use `try_init()`
instead of `init()` so a failure to do the setup results in an
`Result::Err(_)` we can throw away, instead of a panic.