pebble: use owned types for some config
In future commits we will template a config per-test, which means we
can't use the global static config, and will need to construct some
values at runtime (e.g. bind addresses/ports).

This commit switches some Config fields to owned types and makes
`default_config()` a function instead of a static. Notably this
simplifies the handling/serialization of profiles as an added benefit.