pebble: manage challenge test server config
Previously we were only managing the `pebble` configuration, and using
implicit defaults for the `pebble-challtestsrv` configuration.

This commit adds a `EnvironmentConfig` that holds configuration
for both the `challtestsrv` (just two ports presently) and the
`PebbleConfig` (as it was before).

The logic for spawning both subprocesses is updated to use the new
configuration, and to disable some unused features like DoH or HTTP-01
challenges with a redirect from :80 to :443.