README: drop mention of testing limitations
Other folks have reported[0] success with HTTP-01 challenges. Beyond
that there are now integration tests against Pebble for all challenge
types. The Pebble test server was written to be distinct enough from
Let's Encrypts production/staging ACME server to be a good stand in for
general RFC 8555/ACME compatibility and if it works with Pebble it
should work with any standards-compliant CA.

[0]: https://www.shuttle.dev/blog/2025/02/06/provisioning-tls-certificates-with-acme-in-rust