README: more specificity for account key limitation
This limitation is specific to the ACME account key. Users can of course
issue RSA certificates using instant-acme without encountering this
limitation.

Additionally, specify that the account keys must be on the P-256 curve.
Even when using `Acccount::from_parts()` with a pre-existing PKCS8 DER
encoded ECDSA private key it must be P-256.