properly handle potential of authz reuse
Previously both the provision example and the Pebble integration test
logic created a new order for a set of input identifiers, iterated the
authorizations for the resulting order, and then pulled out the
identifier/challenge for each **pending** status authz. After POSTing
each authz's challenge URL the associated challenge identifiers were
used for the CSR sent to finalize the order.

Since the logic above only processes pending authz's it would skip any
that are already **valid** (e.g. from previous orders). This can end up
producing a CSR that has fewer names than the initial order, which
provokes an error during finalization when the CA rejects the CSR.

The solution is simple: don't tie the CSR identifiers to the
challenges at all, use the same list of identifiers as was used to
create the order. Everything else remains the same.