implement replacement orders
By optionally specifying a `CertificateIdentifier` when
constructing a `NewOrder` it's possible to indicate to the ACME CA that
the order is a replacement for a pre-existing certificate that was
previously ordered.

When accepted, an order created in this way will echo back a `replaces`
field with the serialized `CertificateIdentifier` the order replaces.